Move license server code from shared to backend
