+++ packages/back-end/src/enterprise/index.ts
@@ -1,0 +1,1 @@
+export * from "./licenseUtil";
