+++ packages/back-end/src/enterprise/public-key.ts
@@ -1,0 +1,15 @@
+export const LICENSE_PUBLIC_KEY = `-----BEGIN PUBLIC KEY-----
+MIICIjANBgkqhkiG9w0BAQEFAAOCAg8AMIICCgKCAgEArAnMeBofxSrfmkXwytQ3
++VN1L4Ud5zk1xloV8qgagccLN7DWnazgx52qMv4g9eM5M7b8W43acMlb3dFlTp/c
+dBjk/BrjEdOYAvIEG4uTLYiiA1TfbvvcmagGTB+4BV0C8kpigEp6XDgXGRkZmyu+
+EUf8KYqB3/vXdI80e1czUUDlkgllrZn4LzKNoKIViRAMumpiXD7dJFpHcRH+rfF1
+3GmMN0G5nnSYuukjVPQlsAQrz3h77LgCBqLIbcWrCSb88rtXrgI1CMDPdTtHFNyn
+svaH2PK0XPV+3ZS+A+TgP3POfElzkz5Hwrxb6ZTljWpfnmH/Vj/I0huW7jhxYQR4
+lx18ki62nnmNilocyMsbQA6L4OPX/sR9cdHYsAUMykF5rB7RqweR5EQb8FYW1l1l
+TxEKWPaiuY9zBlQn8JzCRZ36kSVOhsd5THrmWaCFTIxHhWUEGeWPvXWBsVerLvHT
+phI91z9zKUogKV105C6s1LxhoKBtlYD07b6dK8QoOkhOi59KhtiLZd5zUgDZUnEX
+wr1TqrKxAQABIyApHzKqgrKtmksw4bsX8sn5ahUgHBdIVH1mNduuLKvPidYQ69PQ
+6iaqBacQ+ZheYxcSaH3BsSXFzDPOLQ2Kj9JBDxEmea2uP0GPUaTGoIwITsVvXCA4
+bKoIzRx9nfcEorfdXWTgNIkCAwEAAQ==
+-----END PUBLIC KEY-----
+`;
