+++ packages/back-end/src/jobs/updateLicense.ts
@@ -5,7 +5,6 @@
  * void within a week if something is not done to unblock the connection.
  */
 import Agenda from "agenda";
-import { licenseInit } from "shared/enterprise";
 import { getSelfHostedOrganization } from "back-end/src/models/OrganizationModel";
 import { trackJob } from "back-end/src/services/otel";
 import { IS_CLOUD } from "back-end/src/util/secrets";
@@ -13,6 +12,7 @@ import {
   getLicenseMetaData,
   getUserCodesForOrg,
 } from "back-end/src/services/licenseData";
+import { licenseInit } from "back-end/src/enterprise";
 
 const UPDATE_LICENSES_JOB_NAME = "updateLicenses";
