+++ packages/back-end/src/scim/middleware/scimMiddleware.ts
@@ -1,7 +1,7 @@
 import { Response, NextFunction } from "express";
-import { orgHasPremiumFeature } from "shared/enterprise";
 import { usingOpenId } from "back-end/src/services/auth";
 import { BaseScimRequest, ScimError } from "back-end/types/scim";
+import { orgHasPremiumFeature } from "back-end/src/enterprise";
 
 export default function scimMiddleware(
   req: BaseScimRequest,
