+++ packages/back-end/src/services/experimentNotifications.ts
@@ -4,7 +4,7 @@ import {
   getMetricResultStatus,
   getExperimentResultStatus,
 } from "shared/experiments";
-import { orgHasPremiumFeature } from "shared/enterprise";
+import { orgHasPremiumFeature } from "back-end/src/enterprise";
 import { StatsEngine } from "back-end/types/stats";
 import { Context } from "back-end/src/models/BaseModel";
 import { createEvent, CreateEventData } from "back-end/src/models/EventModel";
