+++ packages/back-end/src/util/migrations.ts
@@ -6,7 +6,6 @@ import {
   DEFAULT_STATS_ENGINE,
 } from "shared/constants";
 import { RESERVED_ROLE_IDS, getDefaultRole } from "shared/permissions";
-import { accountFeatures, getAccountPlan } from "shared/enterprise";
 import { omit } from "lodash";
 import { SavedGroupInterface } from "shared/src/types";
 import { v4 as uuidv4 } from "uuid";
@@ -43,6 +42,7 @@ import {
 } from "back-end/types/experiment-snapshot";
 import { getEnvironments } from "back-end/src/services/organizations";
 import { LegacySavedGroupInterface } from "back-end/types/saved-group";
+import { accountFeatures, getAccountPlan } from "back-end/src/enterprise";
 import { DEFAULT_CONVERSION_WINDOW_HOURS } from "./secrets";
 
 function roundVariationWeight(num: number): number {
