+++ packages/shared/src/enterprise/index.ts
@@ -1,4 +1,4 @@
-export * from "./licenseUtil";
+export * from "./license-consts";
 export * from "./sso";
 export * from "./pipeline";
 export * from "./power";
